changed styling in lobby and moved files in /components/
