+++ app/components/game/ClueHistory.tsx
