+++ app/components/game/ConfirmPenaltyCardRevealModal.tsx
