+++ app/components/game/QuitGameModal.tsx
