+++ app/components/game/ReportConfirmationModal.tsx
