+++ app/styles/teamTable.module.css
@@ -1,6 +1,7 @@
 .teamBox {
   min-width: 0;
   border-radius: 8px;
+  border: 2px solid rgba(255, 255, 255, 0.65);
   padding: 12px;
 }
 
@@ -15,6 +16,7 @@
 }
 
 .teamTitle {
+  -webkit-text-stroke: 0.25px black;
   text-align: center;
   font-size: clamp(16px, 2.5vw, 20px);
   margin: 0 0 8px;
